Extract layout JSON serializer and add schema versioned document
